Fix msvc2015 compilation warning: ..\src\SMESH\SMESH_MeshEditor.cpp(6968): warning C4838: conversion from 'unsigned __int64' to 'int' requires a narrowing conversion

+++ src/SMESH/SMESH_MeshEditor.cpp
@@ -6966,7 +6966,7 @@ SMESH_MeshEditor::SewFreeBorder (const SMDS_MeshNode* theBordFirstNode,
   // ============================
 
   int nbNodes[]  = { nSide[0].size(), nSide[1].size() };
-  int maxNbNodes = Max( nbNodes[0], nbNodes[1] );
+  int maxNbNodes = (int)Max( nbNodes[0], nbNodes[1] );
 
   TListOfListOfNodes nodeGroupsToMerge;
   if ( nbNodes[0] == nbNodes[1] ||
